requirements/development.txt
-r base.txt

coverage==4.4.1
coveralls==1.2.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.2.2
requests-mock==1.3.0
